minor renamings
