+++ qiskit/aqua/algorithms/single_sample/amplitude_estimation/__init__.py
