+++ qiskit/aqua/algorithms/single_sample/amplitude_estimation/amplitude_estimation.py
