+++ qiskit/aqua/algorithms/single_sample/amplitude_estimation/s_0_reflection.py
