+++ qiskit/aqua/algorithms/single_sample/amplitude_estimation/s_psi_0_reflection.py
