+++ qiskit/aqua/algorithms/single_sample/bernstein_vazirani/__init__.py
