+++ qiskit/aqua/algorithms/single_sample/bernstein_vazirani/bernstein_vazirani.py
