+++ qiskit/aqua/algorithms/single_sample/deutsch_josza/__init__.py
