+++ qiskit/aqua/algorithms/single_sample/deutsch_josza/deutsch_jozsa.py
